Straighten up the helmet situation
- `MarketplaceHelmet` is now *just* `Helmet`
- Removed `suffix` and `image` props because they weren't customised anywhere.
- `otherTags` is now a `ReactNode` which makes it easier to use (less complexity and custom mappings).
- 48 lines instead of 60.
