+++ package.json
@@ -179,6 +179,7 @@
         "@types/react-autocomplete": "^1.8.6",
         "@types/react-avatar-editor": "^13.0.0",
         "@types/react-dom": "^18.2.4",
+        "@types/react-helmet": "^6.1.6",
         "@types/react-notification-system": "^0.2.42",
         "@types/react-router-dom": "^5.3.3",
         "@types/styled-components": "^5.1.26",
