+++ src/pages/NetworkOverviewPage.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import styles from '~/pages/ProjectListingPage.pcss'
 import Layout, { PageContainer } from '~/shared/components/Layout'
-import { NetworkHelmet } from '~/shared/components/Helmet'
+import { NetworkHelmet } from '~/components/Helmet'
 import {
     WhiteBox,
     WhiteBoxPaddingStyles,
