+++ src/pages/OperatorsPage/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { toaster } from 'toasterhea'
 import Button from '~/shared/components/Button'
-import { MarketplaceHelmet } from '~/shared/components/Helmet'
+import Helmet from '~/components/Helmet'
 import Layout from '~/shared/components/Layout'
 import LoadingIndicator from '~/shared/components/LoadingIndicator'
 import { PageWrap } from '~/shared/components/PageWrap'
@@ -25,7 +25,7 @@ const joinSponsorshipModal = toaster(JoinSponsorshipModal, Layer.Modal)
 export default function OperatorsPage() {
     return (
         <Layout gray>
-            <MarketplaceHelmet title="Streams" />
+            <Helmet title="Streams" />
             <LoadingIndicator />
             <PageWrap>
                 <Button
