+++ src/pages/SingleSponsorshipPage.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { useParams } from 'react-router-dom'
 import { useQuery } from '@tanstack/react-query'
 import styles from '~/pages/ProjectListingPage.pcss'
-import { NetworkHelmet } from '~/shared/components/Helmet'
+import { NetworkHelmet } from '~/components/Helmet'
 import Layout, { PageContainer } from '~/shared/components/Layout'
 import { useSponsorship } from '~/hooks/useSponsorship'
 import { NoData } from '~/shared/components/NoData'
